restructured tests and included entity tests
