+++ src/test/kotlin/org/olaven/enterprise/trees/controller/ControllerTestBase.kt
@@ -4,41 +4,19 @@ import com.github.javafaker.Faker
 import io.restassured.RestAssured
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.extension.ExtendWith
-import org.olaven.enterprise.trees.DatabaseReset
+import org.olaven.enterprise.trees.TestBase
 import org.olaven.enterprise.trees.TreeApplication
-import org.olaven.enterprise.trees.dto.LocationDTO
-import org.olaven.enterprise.trees.dto.PlantDto
-import org.olaven.enterprise.trees.entity.LocationEntity
-import org.olaven.enterprise.trees.entity.PlantEntity
-import org.olaven.enterprise.trees.repository.LocationRepository
-import org.olaven.enterprise.trees.repository.PlantRepository
-import org.olaven.enterprise.trees.transformer.LocationTransformer
-import org.olaven.enterprise.trees.transformer.PlantTransformer
-import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.boot.web.server.LocalServerPort
 import org.springframework.test.context.junit.jupiter.SpringExtension
 
 @ExtendWith(SpringExtension::class)
 @SpringBootTest(classes = [ TreeApplication::class ], webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-abstract class ControllerTestBase {
+abstract class ControllerTestBase: TestBase() {
 
     @LocalServerPort
     protected var port = 0
 
-    @Autowired
-    private lateinit var databaseReset: DatabaseReset
-    @Autowired
-    protected lateinit var locationRepository: LocationRepository
-    @Autowired
-    protected lateinit var plantRepository: PlantRepository
-
-    @Autowired
-    private lateinit var locationTransformer: LocationTransformer
-    @Autowired
-    private lateinit var plantTransformer: PlantTransformer
-
-    protected val faker = Faker()
 
     @BeforeEach
     fun init() {
@@ -49,38 +27,5 @@ abstract class ControllerTestBase {
         RestAssured.enableLoggingOfRequestAndResponseIfValidationFails()
 
         databaseReset.reset();
-    }
-
-    protected fun persistLocation(dto: LocationDTO): LocationEntity {
-
-        val entity = locationTransformer.toEntity(dto)
-        return locationRepository.save(entity)
-    }
-
-    protected fun persistPlant(dto: PlantDto): PlantEntity {
-
-        val entity = plantTransformer.toEntity(dto)
-        return plantRepository.save(entity)
-    }
-
-    protected fun getLocationDTO(): LocationDTO {
-
-        val x = faker.random().nextDouble()
-        val y = faker.random().nextDouble()
-
-        return LocationDTO(x, y, null)
-    }
-
-    protected fun getPlantDTO(location: LocationEntity? = null): PlantDto {
-
-        val name = faker.funnyName().name()
-        val description = faker.lorem().paragraph()
-        val height = faker.number().randomDouble(2, 1, 100)
-        val age = faker.number().numberBetween(4, 20)
-
-        val actualLocation = location ?: persistLocation(getLocationDTO())
-        val locationDTO = locationTransformer.toDTO(actualLocation)
-
-        return PlantDto(name, description, height, age, locationDTO)
     }
 }
