+++ src/test/kotlin/org/olaven/enterprise/trees/entity/EntityTestBase.kt
@@ -1,0 +1,5 @@
+package org.olaven.enterprise.trees.entity
+
+import org.olaven.enterprise.trees.TestBase
+
+open class EntityTestBase: TestBase()
